Moved the run query to runs/$runId, instead of /tasks
